Rename folders to `ranges` and `operations`
