+++ src/operations/AbstractOperation.ts
