+++ src/ranges/LinesRange.ts
