+++ src/ranges/MarkedTextRange.ts
