+++ src/ranges/TextRange.ts
